refactor for json part way thorugh
