+++ src/main/java/training/supportbank/JsonReader.java
@@ -1,0 +1,13 @@
+package training.supportbank;
+import com.google.code.gson;
+import com.google.gson.GsonBuilder;
+
+import java.time.LocalDate;
+
+public class JsonReader {
+//    GsonBuilder gsonBuilder = new GsonBuilder();
+//    gsonBuilder.registerTypeAdapter(LocalDate .class,(JsonDeserializer<LocalDate>)(jsonElement,type,jsonDeserializationContext)->
+//            // Convert jsonElement to a LocalDate here...
+//            );
+//    Gson gson = gsonBuilder.create();
+}
